pyproject.toml
[project]
name = "global-news-mcp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "fastmcp>=2.13.0.2",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pygithub>=2.8.1",
    "python-dateutil>=2.9.0.post0",
]

[[tool.uv.index]]
url = "http://mirrors.cloud.aliyuncs.com/pypi/simple"
default = true
